+++ CHANGELOG.md
@@ -2,5 +2,9 @@
 
 All `sniffnet-agent` releases with the relative changes are documented in this file.
 
+## UNRELEASED
+- Support IGMP flows
+- Support VLAN-tagged flows ([#24](https://github.com/GyulyVGC/sniffnet-agent/pull/24))
+
 ## [0.1.0] - 2026-09-02
 - Initial release of `sniffnet-agent`, a lightweight network flows exporter compatible with Sniffnet